enhancement(prometheus_scrape source): update bucket size to u64 (#13318)
* Update count to u64


* Clarify parse_value as equivalent to Go's ParseFloat.


* Update ddsketch to error on too big buckets


* Formatting


* Clippy


* Update fixtures for `roundtrip_native_fixtures`.


* Add new test to ensure old versions of event proto still deserialize


* Add unit test for out of range buckets


* Check precision upto 2^54 for parsing.

+++ lib/vector-core/src/event/lua/metric.rs
@@ -207,7 +207,7 @@ impl<'a> FromLua<'a> for Metric {
         } else if let Some(aggregated_histogram) =
             table.raw_get::<_, Option<LuaTable>>("aggregated_histogram")?
         {
-            let counts: Vec<u32> = aggregated_histogram.raw_get("counts")?;
+            let counts: Vec<u64> = aggregated_histogram.raw_get("counts")?;
             let buckets: Vec<f64> = aggregated_histogram.raw_get("buckets")?;
             let count = counts.iter().sum();
             MetricValue::AggregatedHistogram {
